api/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
pandas==2.2.3
numpy==2.1.3
scikit-learn==1.6.0
eodhd==1.0.32
python-dotenv==1.0.1
pydantic==2.10.4
yfinance==0.2.36
requests==2.32.3
beautifulsoup4==4.12.3
tradingview-screener==3.0.0
tvDatafeed==2.1.6
